+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 7.99
 -----
-
+*   Updates
+    *   Improve episode rows swipe and multi-select animations.
+        ([#4500](https://github.com/Automattic/pocket-casts-android/pull/4500))
 
 7.98
 -----